rename comp
