+++ packages/elements/src/index.ts
@@ -21,8 +21,7 @@ import './photon-state-input';
 import './photon-patient-dialog';
 import './photon-update-patient-dialog';
 import './photon-medication-search';
-import './photon-advanced-medication-search';
-import './photon-advanced-medication-search-dialog';
+import './photon-add-medication-history-dialog';
 import './photon-med-history';
 import './photon-radio-card';
 import './photon-radio-group';
