+++ packages/elements/src/photon-add-medication-history-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from './photon-add-medication-history-dialog-component';
