--- packages/elements/src/photon-advanced-medication-search-dialog/index.ts
@@ -1,1 +1,0 @@
-export * from './photon-advanced-medication-search-dialog-component';
