--- packages/elements/src/photon-advanced-medication-search/style.css
@@ -1,64 +1,0 @@
-.invalid::part(base) {
-  @apply border-red-400;
-  @apply border-2;
-}
-
-.input::part(base) {
-  border-radius: var(--photon-border-radius, 0.5rem);
-  @apply hover:cursor-pointer;
-}
-
-.input::part(input) {
-  @apply overflow-hidden;
-  @apply overflow-ellipsis;
-  @apply whitespace-nowrap;
-}
-
-.disabled::part(base) {
-  @apply hover:cursor-not-allowed;
-}
-
-.selected::part(base) {
-  @apply bg-[#3182CE];
-  @apply text-white;
-}
-
-.default::part(base) {
-  @apply hover:bg-gray-200;
-  @apply hover:text-gray-800;
-}
-
-.default::part(label) {
-  @apply w-full;
-  @apply overflow-hidden;
-}
-
-.group-member::part(base) {
-  @apply pl-4;
-  @apply font-medium;
-}
-
-.group::part(base) {
-  background: #d1e9ff;
-  @apply py-1;
-  @apply text-gray-800;
-  font-size: 0.75rem;
-  font-weight: 600;
-}
-
-.dropdown::part(panel) {
-  border-radius: var(--photon-border-radius, 0.5rem);
-}
-
-.dropdown-container {
-  border-radius: var(--photon-border-radius, 0.5rem);
-}
-
-.group::part(checked-icon) {
-  display: none;
-  width: 0px;
-}
-.group::part(label) {
-  padding-left: 12px;
-  padding-right: 12px;
-}
