+++ packages/elements/src/photon-multirx-form/components/AddPrescriptionCard.tsx
@@ -79,7 +79,7 @@ export const AddPrescriptionCard = (props: {
   enableOrder: boolean;
 }) => {
   const client = usePhoton();
-  const [medDialogOpen, setMedDialogOpen] = createSignal(false);
+  // const [medDialogOpen, setMedDialogOpen] = createSignal(false);
   const [offCatalog, setOffCatalog] = createSignal<Medication | undefined>(undefined);
   const [dispenseUnit] = createSignal<DispenseUnit | undefined>(undefined);
   const [openDoseCalculator, setOpenDoseCalculator] = createSignal(false);
@@ -240,7 +240,7 @@ export const AddPrescriptionCard = (props: {
         <div
           class="flex flex-col sm:gap-3"
           on:photon-medication-selected={(e: any) => {
-            setMedDialogOpen(false);
+            // setMedDialogOpen(false);
             setOffCatalog(e.detail.medication);
             props.actions.updateFormValue({
               key: 'treatment',
@@ -318,11 +318,11 @@ export const AddPrescriptionCard = (props: {
                   })
                 }
               />
-              <photon-advanced-medication-search-dialog
-                title="Advanced Medication Search"
+              {/* <photon-add-medication-history-dialog
+                title="Add Medication History"
                 open={medDialogOpen()}
                 on:photon-medication-closed={() => setMedDialogOpen(false)}
-              />
+              /> */}
             </div>
           </div>
           <div class="mt-2 sm:mt-0 w-full md:pr-2">
